Added feature - Lumantas

+++ delete.php
@@ -1,13 +1,17 @@
 <?php
 require 'db.php';
-if(confirm("Are you sure you want to delete?")) {
-    // proceed delete
-}
 
-if(isset($_GET['delete_id'])){
+if (isset($_GET['delete_id']) && !empty($_GET['delete_id'])) {
+
     $id = $_GET['delete_id'];
+
     $stmt = $pdo->prepare("DELETE FROM students WHERE id = ?");
     $stmt->execute([$id]);
-}
 
-?>
+    header("Location: index.php");
+    exit();
+
+} else {
+    header("Location: index.php");
+    exit();
+}
